6 size

+++ 0x01-variables_if_else_while/6-size.c
@@ -1,21 +1,21 @@
 #include <stdio.h>
 /**
- * main - A program that prints the size
+ * main - function of mine
  * Return:0 (success)
  */
 int main(void)
 {
 
 
-	printf("Size of a char: %i byte(s)", (unsigned long)sizeof(char));
+	printf("Size of a char: %i byte(s)", sizeof(char));
 
-	printf("Size of an int: %i byte(s)", (unsigned long)sizeof(int));
+	printf("Size of an int: %i byte(s)", sizeof(int));
 
-	printf("Size of a long int: %i byte(s)", (unsigned long)sizeof(long int));
+	printf("Size of a long int: %i byte(s)", sizeof(long int));
 
-	printf("Size of a long long int: %i byte(s)", (unsigned long)sizeof(long long int));
+	printf("Size of a long long int: %i byte(s)", sizeof(long long int));
 
-	printf("Size of a float: %i byte(s)", (unsigned long)sizeof(float));
+	printf("Size of a float: %i byte(s)", sizeof(float));
 	printf("\n");
 
 	return (0);
